Disable the SSH client tests on Windows
The tests haven't worked on Windows since PR #5352 landed and making
them work there now has proved to be quite difficult, either involving
the use of some terrible, unmaintable batch scripting or some major
changes to the tests. Given that the tests only exercise the openssh
based connections (as opposed to gocrypto) and that it's unlikely that
openssh client software will actually be available on a windows client
machine, we don't lose much by not running these test on Windows.

Fixes LP #1584815.
