+++ cmd/juju/commands/ssh_common_test.go
@@ -104,6 +104,28 @@ type SSHCommonSuite struct {
 	binDir        string
 }
 
+// Commands to patch
+var patchedCommands = []string{"ssh", "scp"}
+
+// fakecommand outputs its arguments to stdout for verification
+var fakecommand = `#!/bin/bash
+
+{
+    echo "$@"
+
+    # If a custom known_hosts file was passed, emit the contents of
+    # that too.
+    while (( "$#" )); do
+        if [[ $1 = UserKnownHostsFile* ]]; then
+            IFS=" " read -ra parts <<< $1
+            cat "${parts[1]}"
+            break
+        fi
+        shift
+    done
+}| tee $0.args
+`
+
 func (s *SSHCommonSuite) SetUpTest(c *gc.C) {
 	s.JujuConnSuite.SetUpTest(c)
 	ssh.ClearClientKeys()
